backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
httpx
python-dotenv
tonsdk
TonTools
pytest
